Fix Render deployment: Replace mssql with pg for PostgreSQL migration
