+++ backend/package.json
@@ -4,8 +4,7 @@
   "description": "Games Raffle Backend API",
   "main": "src/server.js",
   "scripts": {
-    "start": "node src/server.js",
-    "postinstall": "npm install --production"
+    "start": "node src/server.js"
   },
   "dependencies": {
     "express": "^4.18.2",
@@ -16,7 +15,7 @@
     "bcrypt": "^5.1.0",
     "jsonwebtoken": "^9.0.2",
     "uuid": "^9.0.0",
-    "mssql": "^9.1.1",
+    "pg": "^8.11.3",
     "nodemailer": "^6.9.4",
     "twilio": "^4.14.0",
     "dotenv": "^16.3.1",
@@ -24,6 +23,6 @@
     "node-cron": "^3.0.2"
   },
   "engines": {
-    "node": ">=18.0.0"
+    "node": ">=18.0.0 <=24.0.0"
   }
 }
