+++ backend/src/config/supabase.js
@@ -2,7 +2,7 @@ const { Pool } = require('pg');
 const config = require('./config');
 const logger = require('../utils/logger');
 
-class SupabaseDatabase {
+class PostgreSQLDatabase {
   constructor() {
     this.pool = null;
     this.isConnected = false;
@@ -10,7 +10,7 @@ class SupabaseDatabase {
 
   async initialize() {
     try {
-      logger.info('Initializing PostgreSQL connection to Supabase...');
+      logger.info('Initializing PostgreSQL database connection...');
       
       if (!config.database.connectionString) {
         throw new Error('Database connection string is not configured');
@@ -28,7 +28,8 @@ class SupabaseDatabase {
 
       // Handle pool events
       this.pool.on('connect', () => {
-        logger.info('New PostgreSQL client connected');
+        logger.info('PostgreSQL client connected');
+        this.isConnected = true;
       });
 
       this.pool.on('error', (err) => {
@@ -40,10 +41,10 @@ class SupabaseDatabase {
       await this.testConnection();
       
       this.isConnected = true;
-      logger.info('PostgreSQL connection established successfully');
+      logger.info('PostgreSQL database connection established successfully');
       
     } catch (error) {
-      logger.error('Failed to initialize PostgreSQL connection:', error);
+      logger.error('Failed to initialize PostgreSQL database connection:', error);
       throw error;
     }
   }
@@ -51,31 +52,110 @@ class SupabaseDatabase {
   async testConnection() {
     try {
       const client = await this.pool.connect();
-      const result = await client.query('SELECT NOW()');
+      const result = await client.query('SELECT 1 as test');
       client.release();
-      logger.info('PostgreSQL connection test successful:', result.rows[0]);
-      return true;
+      
+      if (result.rows[0].test === 1) {
+        logger.info('PostgreSQL database connection test passed');
+        return true;
+      } else {
+        throw new Error('PostgreSQL database connection test failed');
+      }
     } catch (error) {
-      logger.error('PostgreSQL connection test failed:', error);
+      logger.error('PostgreSQL database connection test failed:', error);
       throw error;
     }
   }
 
-  async query(text, params = []) {
+  async query(queryText, params = []) {
     if (!this.pool) {
       throw new Error('Database not initialized. Call initialize() first.');
     }
 
-    const client = await this.pool.connect();
     try {
-      const start = Date.now();
-      const result = await client.query(text, params);
-      const duration = Date.now() - start;
+      // Convert named parameters to positional parameters for PostgreSQL
+      let pgQuery = queryText;
+      let pgParams = [];
       
-      logger.debug('Executed query', { text, duration, rows: result.rowCount });
+      if (typeof params === 'object' && !Array.isArray(params)) {
+        let paramIndex = 1;
+        const paramMap = {};
+        
+        // Replace @paramName with $1, $2, etc.
+        pgQuery = queryText.replace(/@(\w+)/g, (match, paramName) => {
+          if (!paramMap[paramName]) {
+            paramMap[paramName] = paramIndex++;
+            pgParams.push(params[paramName]);
+          }
+          return `$${paramMap[paramName]}`;
+        });
+      } else {
+        pgParams = params;
+      }
+      
+      const result = await this.pool.query(pgQuery, pgParams);
+      
+      // Return normalized result that matches SQL Server format
+      return {
+        rows: result.rows,
+        rowCount: result.rowCount
+      };
+    } catch (error) {
+      logger.error('PostgreSQL database query error:', { query: queryText, params, error: error.message });
+      throw error;
+    }
+  }
+
+  async transaction(callback) {
+    const client = await this.pool.connect();
+    
+    try {
+      await client.query('BEGIN');
+      
+      // Create a transaction context with normalized query method
+      const transactionContext = {
+        query: async (queryText, params = []) => {
+          // Convert named parameters to positional parameters for PostgreSQL
+          let pgQuery = queryText;
+          let pgParams = [];
+          
+          if (typeof params === 'object' && !Array.isArray(params)) {
+            let paramIndex = 1;
+            const paramMap = {};
+            
+            // Replace @paramName with $1, $2, etc.
+            pgQuery = queryText.replace(/@(\w+)/g, (match, paramName) => {
+              if (!paramMap[paramName]) {
+                paramMap[paramName] = paramIndex++;
+                pgParams.push(params[paramName]);
+              }
+              return `$${paramMap[paramName]}`;
+            });
+          } else {
+            pgParams = params;
+          }
+          
+          const result = await client.query(pgQuery, pgParams);
+          return {
+            rows: result.rows,
+            rowCount: result.rowCount
+          };
+        }
+      };
+      
+      const result = await callback(transactionContext);
+      
+      await client.query('COMMIT');
       return result;
     } catch (error) {
-      logger.error('Database query error:', { text, params, error: error.message });
+      logger.error('PostgreSQL database transaction error:', error);
+      
+      try {
+        await client.query('ROLLBACK');
+      } catch (rollbackError) {
+        logger.error('PostgreSQL database transaction rollback error:', rollbackError);
+      }
+      
       throw error;
     } finally {
       client.release();
@@ -91,9 +171,16 @@ class SupabaseDatabase {
       const result = await this.query('SELECT 1 as health_check');
       return result.rows.length > 0 && result.rows[0].health_check === 1;
     } catch (error) {
-      logger.error('Database health check failed:', error);
+      logger.error('PostgreSQL database health check failed:', error);
       return false;
     }
+  }
+
+  getPool() {
+    if (!this.pool) {
+      throw new Error('Database not connected. Call initialize() first.');
+    }
+    return this.pool;
   }
 
   async close() {
@@ -102,32 +189,104 @@ class SupabaseDatabase {
         await this.pool.end();
         this.pool = null;
         this.isConnected = false;
-        logger.info('PostgreSQL connection pool closed');
+        logger.info('PostgreSQL database connection closed');
       }
     } catch (error) {
-      logger.error('Error closing PostgreSQL connection:', error);
+      logger.error('Error closing PostgreSQL database connection:', error);
       throw error;
     }
   }
 
-  // Helper method to handle transactions
-  async transaction(callback) {
-    const client = await this.pool.connect();
-    try {
-      await client.query('BEGIN');
-      const result = await callback(client);
-      await client.query('COMMIT');
-      return result;
-    } catch (error) {
-      await client.query('ROLLBACK');
-      throw error;
-    } finally {
-      client.release();
-    }
+  // Helper methods for common queries (PostgreSQL versions)
+  async findById(tableName, id, idColumn = 'id') {
+    const query = `SELECT * FROM ${tableName} WHERE ${idColumn} = @id`;
+    const result = await this.query(query, { id });
+    return result.rows[0] || null;
+  }
+
+  async findAll(tableName, whereClause = '', params = {}) {
+    const query = `SELECT * FROM ${tableName} ${whereClause ? 'WHERE ' + whereClause : ''}`;
+    const result = await this.query(query, params);
+    return result.rows;
+  }
+
+  async insert(tableName, data) {
+    const columns = Object.keys(data);
+    const values = columns.map(col => `@${col}`);
+    
+    const query = `
+      INSERT INTO ${tableName} (${columns.join(', ')}) 
+      VALUES (${values.join(', ')})
+      RETURNING *
+    `;
+    
+    const result = await this.query(query, data);
+    return result.rows[0];
+  }
+
+  async update(tableName, id, data, idColumn = 'id') {
+    const columns = Object.keys(data);
+    const setClause = columns.map(col => `${col} = @${col}`).join(', ');
+    
+    const query = `
+      UPDATE ${tableName} 
+      SET ${setClause}
+      WHERE ${idColumn} = @id
+      RETURNING *
+    `;
+    
+    const params = { ...data, id };
+    const result = await this.query(query, params);
+    return result.rows[0];
+  }
+
+  async delete(tableName, id, idColumn = 'id') {
+    const query = `DELETE FROM ${tableName} WHERE ${idColumn} = @id`;
+    const result = await this.query(query, { id });
+    return result.rowCount > 0;
+  }
+
+  // Pagination helper (PostgreSQL version)
+  async paginate(tableName, options = {}) {
+    const {
+      page = 1,
+      limit = 10,
+      orderBy = 'created_at',
+      orderDirection = 'DESC',
+      whereClause = '',
+      params = {}
+    } = options;
+    
+    const offset = (page - 1) * limit;
+    
+    const countQuery = `SELECT COUNT(*) as total FROM ${tableName} ${whereClause ? 'WHERE ' + whereClause : ''}`;
+    const dataQuery = `
+      SELECT * FROM ${tableName} 
+      ${whereClause ? 'WHERE ' + whereClause : ''}
+      ORDER BY ${orderBy} ${orderDirection}
+      LIMIT @limit OFFSET @offset
+    `;
+    
+    const [countResult, dataResult] = await Promise.all([
+      this.query(countQuery, params),
+      this.query(dataQuery, { ...params, offset, limit })
+    ]);
+    
+    const total = parseInt(countResult.rows[0].total);
+    const totalPages = Math.ceil(total / limit);
+    
+    return {
+      data: dataResult.rows,
+      pagination: {
+        page,
+        limit,
+        total,
+        totalPages,
+        hasNext: page < totalPages,
+        hasPrev: page > 1
+      }
+    };
   }
 }
 
-// Create singleton instance
-const supabaseDb = new SupabaseDatabase();
-
-module.exports = supabaseDb;
+module.exports = PostgreSQLDatabase;
